feat: remove add device from home screen
