+++ apps/enmeshed/lib/account/home/widgets/home_widgets.dart
@@ -1,2 +1,2 @@
-export 'add_contact_or_device_container.dart';
+export 'add_contact.dart';
 export 'complete_profile_container.dart';
